requirements.txt
astroid==2.3.3
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.4.4
python-telegram-bot==12.5.1
six==1.14.0
tornado==6.0.4
typed-ast==1.4.1
wrapt==1.11.2
